+++ SlimHUD/Services/DisplayManager.swift
@@ -2,8 +2,7 @@
 //  DisplayManager.swift
 //  SlimHUD
 //
-//  Created by Alex Perathoner on 19/12/2022.
-//  Copyright © 2022 Alex Perathoner. All rights reserved.
+//  Created by Alex Perathoner on 24/12/22.
 //
 
 import Foundation
@@ -11,46 +10,53 @@ import Cocoa
 
 class DisplayManager {
     private init() {}
-    
+
     private static var useM1DisplayBrightnessMethod = false
-    
-    static func getDisplayBrightness() -> Float {
-        if(useM1DisplayBrightnessMethod) {
-            return getM1DisplayBrightness()
-        } else {
+
+    private static var method = SensorMethod.standard
+
+    static func getDisplayBrightness() throws -> Float {
+        switch DisplayManager.method {
+        case .standard:
             do {
                 return try getStandardDisplayBrightness()
             } catch {
-                DisplayManager.useM1DisplayBrightnessMethod = true
-                return getM1DisplayBrightness()
+                method = .m1
             }
+        case .m1:
+            do {
+                return try getM1DisplayBrightness()
+            } catch {
+                method = .allFailed
+            }
+        case .allFailed:
+            throw SensorError.Display.notFound
         }
+        return try getDisplayBrightness()
     }
-    
+
     private static func getStandardDisplayBrightness() throws -> Float {
         var brightness: float_t = 1
         let service = IOServiceGetMatchingService(kIOMasterPortDefault, IOServiceMatching("IODisplayConnect"))
         defer {
             IOObjectRelease(service)
         }
-        
+
         let result = IODisplayGetFloatParameter(service, 0, kIODisplayBrightnessKey as CFString, &brightness)
-        if(result != kIOReturnSuccess) {
-            throw SensorError.displayBrightnessFailure
+        if result != kIOReturnSuccess {
+            throw SensorError.Display.notStandard
         }
         return brightness
     }
-    private static func getM1DisplayBrightness() -> Float {
+    private static func getM1DisplayBrightness() throws -> Float {
         let task = Process()
         task.launchPath = "/usr/libexec/corebrightnessdiag"
         task.arguments = ["status-info"]
         let pipe = Pipe()
         task.standardOutput = pipe
-        task.launch()
+        try task.run()
         let data = pipe.fileHandleForReading.readDataToEndOfFile()
         task.waitUntilExit()
-
-        var displayBrightness: Float?
 
         if let plist = try? PropertyListSerialization.propertyList(from: data, options: [], format: nil) as? NSDictionary {
             if let displays = plist["CBDisplays"] as? [String: [String: Any]] {
@@ -58,15 +64,14 @@ class DisplayManager {
                     if let displayInfo = display["Display"] as? [String: Any],
                         displayInfo["DisplayServicesIsBuiltInDisplay"] as? Bool == true,
                         let brightness = displayInfo["DisplayServicesBrightness"] as? Float {
-                            displayBrightness = brightness
-                            break
+                            return brightness
                     }
                 }
             }
         }
-        return displayBrightness ?? 0 // todo add throws, handle outside
+        throw SensorError.Display.notSilicon
     }
-    
+
     /* Note the difference between NSScreen.main and NSScreen.screens[0]:
      * NSScreen.main is the "key" screen, where the currently frontmost window resides.
      * NSScreen.screens[0] is the screen which has a menu bar, and is chosen in the Preferences > monitor settings
@@ -105,7 +110,7 @@ class DisplayManager {
 
     static func getDockPosition() -> Position {
         var dockPosition: Position = .bottom
-        if let rawPosition = UserDefaults.standard.persistentDomain(forName: "com.apple.dock")!["orientation"] as? String {
+        if let rawPosition = UserDefaults.standard.persistentDomain(forName: "com.apple.dock")?["orientation"] as? String {
             if let actualPosition = Position(rawValue: rawPosition) {
                 dockPosition = actualPosition
             } else {