+++ SlimHUD/Services/UserDefaultsManager.swift
@@ -2,8 +2,7 @@
 //  UserDefaultsManager.swift
 //  SlimHUD
 //
-//  Created by Alex Perathoner on 19/12/2022.
-//  Copyright © 2022 Alex Perathoner. All rights reserved.
+//  Created by Alex Perathoner on 24/12/22.
 //
 
 import Foundation
@@ -13,6 +12,7 @@ class UserDefaultsManager {
     static func getItem<T>(for key: String, defaultValue: T) -> T where T: NSCoding, T: NSObject {
         do {
             guard let data = UserDefaults.standard.object(forKey: key) as? Data,
+                  // swiftlint:disable:next force_cast
                   let item = try NSKeyedUnarchiver.unarchiveTopLevelObjectWithData(data) as! T? else {
                 return defaultValue
             }
@@ -23,11 +23,16 @@ class UserDefaultsManager {
         }
     }
 
-    static func getBool(for key: String) -> Bool {
-        return UserDefaults.standard.bool(forKey: key)
+    static func getBool(for key: String, defaultValue: Bool) -> Bool {
+        if UserDefaults.standard.valueExists(forKey: key) {
+            return UserDefaults.standard.bool(forKey: key)
+        } else {
+            return defaultValue
+        }
     }
 
     static func getArr<T>(for key: String, defaultValue: [T]) -> [T] {
+        // swiftlint:disable:next force_cast
         return UserDefaults.standard.array(forKey: key) as! [T]? ?? defaultValue
     }
 
@@ -35,12 +40,12 @@ class UserDefaultsManager {
         return UserDefaults.standard.string(forKey: key) ?? defaultValue
     }
 
-    static func getString(for key: String) -> String? {
-        return UserDefaults.standard.string(forKey: key)
-    }
-
-    static func getInt(for key: String) -> Int {
-        return UserDefaults.standard.integer(forKey: key)
+    static func getInt(for key: String, defaultValue: Int) -> Int {
+        if UserDefaults.standard.valueExists(forKey: key) {
+            return UserDefaults.standard.integer(forKey: key)
+        } else {
+            return defaultValue
+        }
     }
 
     static func setItem<T>(_ item: T, for key: String) {