+++ SlimHUD/Support/Errors/SensorError.swift
@@ -2,15 +2,20 @@
 //  SensorError.swift
 //  SlimHUD
 //
-//  Created by Alex Perathoner on 20/12/22.
-//  Copyright © 2022 Alex Perathoner. All rights reserved.
+//  Created by Alex Perathoner on 24/12/22.
 //
 
 import Foundation
 
 enum SensorError: Error {
-    case displayBrightnessFailure
-    case keyboardBrightnessFailure
-    case m1DisplayBrightnessFailure
-    case m1KeyboardBrightnessFailure
+    enum Display: Error {
+        case notFound
+        case notSilicon
+        case notStandard
+    }
+    enum Keyboard: Error {
+        case notFound
+        case notSilicon
+        case notStandard
+    }
 }