+++ Release_Notes.md
@@ -1,5 +1,11 @@
-# 1.3.8 - Fixes
+# 1.4.0 - Fixes
 * Fix bug: hud too near to the screen edge
 * Hiding MacOS' default HUD without modifying system configuration files.
 * Fix bug: implemented new ways to retrieve brightness
 * Fix bug: keyboard backlight hud working again
+* Update to Sparkle 2
+* Change in signing certificates: please update manually to the [latest release](https://github.com/AlexPerathoner/SlimHUD/releases/download/v1.4.0/SlimHUD.zip)!
+* Disabling keyboard and display brightness bars until screen settings changes if it wasn't possible to retrieve their values
+
+# 1.3.7 - Added support to macOS 10.13
+* Added support to macOS 10.13